[skipruntime] Split out `server` from `core`.
