+++ package.json
@@ -7,6 +7,7 @@
     "skjson/ts",
     "skipruntime-ts/client",
     "skipruntime-ts/core",
+    "skipruntime-ts/server",
     "skipruntime-ts/examples",
     "sql/ts",
     "sql/ts/tests"
