+++ skipruntime-ts/core/src/skipruntime_errors.ts
@@ -1,0 +1,1 @@
+export class UnknownCollectionError extends Error {}
