+++ skipruntime-ts/examples/package.json
@@ -12,6 +12,7 @@
   "dependencies": {
     "sqlite3": "^5.1.7",
     "@skipruntime/core": "^1.0.0",
+    "@skipruntime/server": "^1.0.0",
     "@skipruntime/client": "^1.0.0"
   }
 }
