+++ skipruntime-ts/server/package.json
@@ -1,0 +1,27 @@
+{
+  "name": "@skipruntime/server",
+  "version": "1.0.0",
+  "type": "module",
+  "main": "./dist/server.js",
+  "types": "./dist/server.d.ts",
+  "exports": {
+    ".": {
+      "default": "./dist/server.js",
+      "types": "./dist/server.d.ts"
+    }
+  },
+  "scripts": {
+    "build": "tsc",
+    "clean": "rm -rf dist",
+    "lint": "eslint"
+  },
+  "devDependencies": {
+    "@types/express": "^4.17.21",
+    "@types/ws": "^8.5.12"
+  },
+  "dependencies": {
+    "express": "^4.21.0",
+    "@skipruntime/client": "^1.0.0",
+    "ws": "^8.18.0"
+  }
+}
