+++ skipruntime-ts/server/src/replication.ts
@@ -1,21 +1,6 @@
-import type { SkipService } from "../skipruntime_service.js";
-import { createSKStore as CreateSKStore } from "../skip-runtime.js";
-import type {
-  Entry,
-  JSONObject,
-  TJSON,
-  SkipRuntime,
-} from "../skipruntime_api.js";
-import express from "express";
-
-import { runService as initService } from "../skipruntime_runner.js";
-import { UnknownCollectionError } from "./skipruntime_impl.js";
-
-import { WebSocket, WebSocketServer, type MessageEvent } from "ws";
-
+import { WebSocket, type WebSocketServer, type MessageEvent } from "ws";
+import type { Entry, TJSON, SkipRuntime } from "@skipruntime/core";
 import { Protocol } from "@skipruntime/client";
-
-import * as http from "http";
 
 class TailingSession {
   private subsessions = new Map<string, bigint>();
@@ -236,139 +221,4 @@ export class ReplicationServer {
       throw error;
     }
   }
-}
-
-export async function runService(
-  service: SkipService,
-  createSKStore: typeof CreateSKStore,
-  port: number = 443,
-): Promise<{ close: () => void }> {
-  const runtime = await initService(service, createSKStore);
-  const app = runRESTServer(runtime);
-  const httpServer = http.createServer();
-  httpServer.on("request", app);
-  const wss = new WebSocketServer({ server: httpServer });
-  const replicationServer = new ReplicationServer(wss, runtime);
-  httpServer.listen(port, () => {
-    console.log(`Reactive service listening on port ${port.toString()}`);
-  });
-
-  return {
-    close: () => {
-      replicationServer.close();
-      httpServer.close();
-    },
-  };
-}
-
-function runRESTServer(runtime: SkipRuntime): express.Express {
-  const app = express();
-  app.use(express.json());
-  app.use(express.urlencoded({ extended: true }));
-  // READS
-  app.head("/v1/:resource", (req, res) => {
-    const resourceName = req.params.resource;
-    const strReactiveAuth = req.headers["x-reactive-auth"] as string;
-    if (!strReactiveAuth) throw new Error("X-Reactive-Auth must be specified.");
-    const reactiveAuth = new Uint8Array(Buffer.from(strReactiveAuth, "base64"));
-    try {
-      const data = runtime.head(
-        resourceName,
-        req.query as JSONObject,
-        reactiveAuth,
-      );
-      res.set("X-Reactive-Response", JSON.stringify(data));
-      res.status(200).json({});
-    } catch (e: unknown) {
-      res.status(500).json(e instanceof Error ? e.message : e);
-    }
-  });
-  app.get("/v1/:resource/:key", (req, res) => {
-    const key = req.params.key;
-    const resourceName = req.params.resource;
-    try {
-      const data = runtime.getOne(resourceName, req.query as JSONObject, key);
-      res.status(200).json(data);
-    } catch (e: unknown) {
-      res.status(500).json(e instanceof Error ? e.message : e);
-    }
-  });
-  app.get("/v1/:resource", (req, res) => {
-    const resourceName = req.params.resource;
-    const strReactiveAuth = req.headers["x-reactive-auth"] as string;
-    const reactiveAuth = strReactiveAuth
-      ? new Uint8Array(Buffer.from(strReactiveAuth, "base64"))
-      : undefined;
-    try {
-      const data = runtime.getAll(
-        resourceName,
-        req.query as Record<string, string>,
-        reactiveAuth,
-      );
-      if (data.reactive) {
-        res.set("X-Reactive-Response", JSON.stringify(data.reactive));
-      }
-      res.status(200).json(data.values);
-    } catch (e: unknown) {
-      res.status(500).json(e instanceof Error ? e.message : e);
-    }
-  });
-  // WRITES
-  app.put("/v1/:collection/:id", (req, res) => {
-    if (!Array.isArray(req.body)) {
-      res.status(400).json(`Bad request body ${JSON.stringify(req.body)}`);
-      return;
-    }
-    const key = req.params.id;
-    const data = req.body as TJSON[];
-    const collectionName = req.params.collection;
-    try {
-      runtime.put(collectionName, key, data);
-      res.status(200).json({});
-    } catch (e: unknown) {
-      if (e instanceof UnknownCollectionError) {
-        res.status(400).json("Bad request");
-      } else {
-        res.status(500).json(e instanceof Error ? e.message : e);
-      }
-    }
-  });
-  app.patch("/v1/:collection", (req, res) => {
-    if (!Array.isArray(req.body)) {
-      res.status(400).json(`Bad request body ${JSON.stringify(req.body)}`);
-      return;
-    }
-    const data = req.body as TJSON[];
-    if (!Array.isArray(data)) {
-      res.status(400).json("Bad request");
-      return;
-    }
-    const collectionName = req.params.collection;
-    try {
-      runtime.patch(collectionName, data as Entry<TJSON, TJSON>[]);
-      res.status(200).json({});
-    } catch (e: unknown) {
-      if (e instanceof UnknownCollectionError) {
-        res.status(400).json("Bad request");
-      } else {
-        res.status(500).json(e instanceof Error ? e.message : e);
-      }
-    }
-  });
-  app.delete("/v1/:collection/:id", (req, res) => {
-    const key = req.params.id;
-    const collectionName = req.params.collection;
-    try {
-      runtime.delete(collectionName, key);
-      res.status(200).json({});
-    } catch (e: unknown) {
-      if (e instanceof UnknownCollectionError) {
-        res.status(400).json("Bad request");
-      } else {
-        res.status(500).json(e instanceof Error ? e.message : e);
-      }
-    }
-  });
-
-  return app;
 }
